Move create maf database out of input to database completely as only relevant for GENIE
