+++ genie/input_to_database.py
@@ -372,66 +372,6 @@ def processfiles(syn, validfiles, center, path_to_genie,
 
     logger.info("ALL DATA STORED IN DATABASE")
 
-# def _create_maf_db(syn, foo):
-#     maf_database_ent = syn.get(maf_database_synid)
-#     print(maf_database_ent)
-#     maf_columns = list(syn.getTableColumns(maf_database_synid))
-#     schema = synapseclient.Schema(
-#         name='Narrow MAF {current_time} Database'.format(
-#             current_time=time.time()),
-#         columns=maf_columns,
-#         parent=process_functions.getDatabaseSynId(
-#             syn, "main",
-#             databaseToSynIdMappingDf=database_synid_mappingdf))
-#     schema.primaryKey = maf_database_ent.primaryKey
-#     new_maf_database = syn.store(schema)
-
-# TODO: Should split this into 3 funcitons
-# so that unit tests are easier to write
-
-
-def create_and_archive_maf_database(syn, database_synid_mappingdf):
-    '''
-    Creates new MAF database and archives the old database in the staging site
-
-    Args:
-        syn: Synapse object
-        databaseToSynIdMappingDf: Database to synapse id mapping dataframe
-
-    Return:
-        Editted database to synapse id mapping dataframe
-    '''
-    maf_database_synid = process_functions.getDatabaseSynId(
-        syn, "vcf2maf", databaseToSynIdMappingDf=database_synid_mappingdf)
-    maf_database_ent = syn.get(maf_database_synid)
-    maf_columns = list(syn.getTableColumns(maf_database_synid))
-    schema = synapseclient.Schema(
-        name='Narrow MAF {current_time} Database'.format(
-            current_time=time.time()),
-        columns=maf_columns,
-        parent=process_functions.getDatabaseSynId(
-            syn, "main", databaseToSynIdMappingDf=database_synid_mappingdf))
-    schema.primaryKey = maf_database_ent.primaryKey
-    new_maf_database = syn.store(schema)
-    # Store in the new database synid
-    database_synid_mappingdf['Id'][
-        database_synid_mappingdf[
-            'Database'] == 'vcf2maf'] = new_maf_database.id
-
-    vcf2maf_mappingdf = database_synid_mappingdf[
-        database_synid_mappingdf['Database'] == 'vcf2maf']
-    # vcf2maf_mappingdf['Id'][0] = newMafDb.id
-    # Update this synid later
-    syn.store(synapseclient.Table("syn12094210", vcf2maf_mappingdf))
-    # Move and archive old mafdatabase (This is the staging synid)
-    maf_database_ent.parentId = "syn7208886"
-    maf_database_ent.name = "ARCHIVED " + maf_database_ent.name
-    syn.store(maf_database_ent)
-    # maf_database_synid = new_maf_database.id
-    # Remove can download permissions from project GENIE team
-    syn.setPermissions(new_maf_database.id, 3326313, [])
-    return(database_synid_mappingdf)
-
 
 def email_duplication_error(syn, duplicated_filesdf):
     '''
